+++ ReleaseNotes.md
@@ -26,6 +26,7 @@ Because of the extension of configuration of transcoding profiles, those produci
 ### FIXES
 
 - Warning in case import function does not return new ids.
+- Config values `<online-content fetch-buffer-size="262144" fetch-buffer-fill-size="0">` are implemented now.
 
 ### Code Improvements
 
@@ -34,7 +35,7 @@ Because of the extension of configuration of transcoding profiles, those produci
 - Upgrade contrib code cxxopts (3.2.1), md5 (2002)
 - Extract code for npupnp and pupnp specific handling
 - Refactoring of UPnP services
-- Cleanup of several todos, FIXMES etc.
+- Cleanup of several todos, FIXMES, code smells etc.
 
 ## v2.2.0